Moving CommonTaskProperties back into task package
