+++ src/main/java/com/libertymutualgroup/herman/aws/ecs/broker/kms/KmsBroker.java
@@ -45,7 +45,7 @@ import com.amazonaws.services.kms.model.TagResourceRequest;
 import com.libertymutualgroup.herman.aws.ecs.EcsPushDefinition;
 import com.libertymutualgroup.herman.aws.ecs.PropertyHandler;
 import com.libertymutualgroup.herman.logging.HermanLogger;
-import com.libertymutualgroup.herman.aws.CommonTaskProperties;
+import com.libertymutualgroup.herman.task.common.CommonTaskProperties;
 import com.libertymutualgroup.herman.util.ConfigurationUtil;
 import com.libertymutualgroup.herman.util.FileUtil;
 import org.slf4j.Logger;
