+++ src/main/java/com/libertymutualgroup/herman/task/cft/CftPushTaskProperties.java
@@ -16,7 +16,7 @@
 package com.libertymutualgroup.herman.task.cft;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.libertymutualgroup.herman.aws.CommonTaskProperties;
+import com.libertymutualgroup.herman.task.common.CommonTaskProperties;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class CftPushTaskProperties extends CommonTaskProperties {
